Fix db column type.

+++ app/database/migrations/2014_12_22_115654_create_prices_table.php
@@ -15,7 +15,7 @@ class CreatePricesTable extends Migration {
 		Schema::create('prices', function(Blueprint $table)
 		{
 			$table->integer('typeID')->primary();
-			$table->float('median_price');
+			$table->integer('median_price');
 			$table->timestamps();
 		});
 	}
